Extract checkout variant picker tests into their own file
