+++ client/my-sites/checkout/composite-checkout/test/composite-checkout.js
@@ -31,10 +31,6 @@ import {
 	mockSetCartEndpoint,
 	mockGetCartEndpointWith,
 	getActivePersonalPlanDataForType,
-	getPersonalPlanForInterval,
-	getBusinessPlanForInterval,
-	getVariantItemTextForInterval,
-	getPlansItemsState,
 	createTestReduxStore,
 	countryList,
 	gSuiteProduct,
@@ -553,159 +549,6 @@ describe( 'CompositeCheckout', () => {
 			expect( screen.getByText( 'Purchase Details' ) ).toBeInTheDocument();
 			expect( page.redirect ).not.toHaveBeenCalled();
 		} );
-	} );
-
-	it.each( [
-		{ activePlan: 'none', cartPlan: 'yearly', expectedVariant: 'monthly' },
-		{ activePlan: 'none', cartPlan: 'yearly', expectedVariant: 'yearly' },
-		{ activePlan: 'none', cartPlan: 'yearly', expectedVariant: 'two-year' },
-		{ activePlan: 'yearly', cartPlan: 'yearly', expectedVariant: 'yearly' },
-		{ activePlan: 'yearly', cartPlan: 'yearly', expectedVariant: 'two-year' },
-		{ activePlan: 'monthly', cartPlan: 'yearly', expectedVariant: 'monthly' },
-		{ activePlan: 'monthly', cartPlan: 'yearly', expectedVariant: 'yearly' },
-		{ activePlan: 'monthly', cartPlan: 'yearly', expectedVariant: 'two-year' },
-		{ activePlan: 'monthly', cartPlan: 'two-year', expectedVariant: 'monthly' },
-		{ activePlan: 'monthly', cartPlan: 'two-year', expectedVariant: 'yearly' },
-		{ activePlan: 'monthly', cartPlan: 'two-year', expectedVariant: 'two-year' },
-	] )(
-		'renders the variant picker with $expectedVariant for a $cartPlan plan when the current plan is $activePlan',
-		async ( { activePlan, cartPlan, expectedVariant } ) => {
-			getPlansBySiteId.mockImplementation( () => ( {
-				data: getActivePersonalPlanDataForType( activePlan ),
-			} ) );
-			const cartChanges = { products: [ getBusinessPlanForInterval( cartPlan ) ] };
-			render( <MyCheckout cartChanges={ cartChanges } />, container );
-			const editOrderButton = await screen.findByLabelText( 'Edit your order' );
-			fireEvent.click( editOrderButton );
-
-			expect(
-				screen.getByText( getVariantItemTextForInterval( expectedVariant ) )
-			).toBeInTheDocument();
-		}
-	);
-
-	it.each( [
-		{ activePlan: 'yearly', cartPlan: 'yearly', expectedVariant: 'monthly' },
-		{ activePlan: 'two-year', cartPlan: 'yearly', expectedVariant: 'monthly' },
-		{ activePlan: 'two-year', cartPlan: 'yearly', expectedVariant: 'yearly' },
-		{ activePlan: 'two-year', cartPlan: 'yearly', expectedVariant: 'two-year' },
-	] )(
-		'renders the variant picker without $expectedVariant for a $cartPlan plan when the current plan is $activePlan',
-		async ( { activePlan, cartPlan, expectedVariant } ) => {
-			getPlansBySiteId.mockImplementation( () => ( {
-				data: getActivePersonalPlanDataForType( activePlan ),
-			} ) );
-			const cartChanges = { products: [ getBusinessPlanForInterval( cartPlan ) ] };
-			render( <MyCheckout cartChanges={ cartChanges } />, container );
-			const editOrderButton = await screen.findByLabelText( 'Edit your order' );
-			fireEvent.click( editOrderButton );
-
-			expect(
-				screen.queryByText( getVariantItemTextForInterval( expectedVariant ) )
-			).not.toBeInTheDocument();
-		}
-	);
-
-	it.each( [
-		{ activePlan: 'none', cartPlan: 'yearly', expectedVariant: 'yearly' },
-		{ activePlan: 'none', cartPlan: 'yearly', expectedVariant: 'two-year' },
-	] )(
-		'renders the $expectedVariant variant with a discount percentage for a $cartPlan plan when the current plan is $activePlan',
-		async ( { activePlan, cartPlan, expectedVariant } ) => {
-			getPlansBySiteId.mockImplementation( () => ( {
-				data: getActivePersonalPlanDataForType( activePlan ),
-			} ) );
-			const cartChanges = { products: [ getBusinessPlanForInterval( cartPlan ) ] };
-			render( <MyCheckout cartChanges={ cartChanges } />, container );
-			const editOrderButton = await screen.findByLabelText( 'Edit your order' );
-			fireEvent.click( editOrderButton );
-
-			const variantItem = screen
-				.getByText( getVariantItemTextForInterval( expectedVariant ) )
-				.closest( 'label' );
-			const lowestVariantItem = variantItem.closest( 'ul' ).querySelector( 'label:first-of-type' );
-			const lowestVariantSlug = lowestVariantItem.closest( 'div' ).querySelector( 'input' ).value;
-			const variantSlug = variantItem.closest( 'div' ).querySelector( 'input' ).value;
-
-			const variantData = getPlansItemsState().find(
-				( plan ) => plan.product_slug === variantSlug
-			);
-			const finalPrice = variantData.raw_price;
-			const variantInterval = variantData.bill_period;
-			const lowestVariantData = getPlansItemsState().find(
-				( plan ) => plan.product_slug === lowestVariantSlug
-			);
-			const lowestVariantPrice = lowestVariantData.raw_price;
-			const lowestVariantInterval = lowestVariantData.bill_period;
-			const intervalsInVariant = Math.round( variantInterval / lowestVariantInterval );
-			const priceBeforeDiscount = lowestVariantPrice * intervalsInVariant;
-
-			const discountPercentage = Math.round( 100 - ( finalPrice / priceBeforeDiscount ) * 100 );
-			expect(
-				within( variantItem ).getByText( `Save ${ discountPercentage }%` )
-			).toBeInTheDocument();
-		}
-	);
-
-	it.each( [ { activePlan: 'none', cartPlan: 'yearly', expectedVariant: 'monthly' } ] )(
-		'renders the $expectedVariant variant without a discount percentage for a $cartPlan plan when the current plan is $activePlan',
-		async ( { activePlan, cartPlan, expectedVariant } ) => {
-			getPlansBySiteId.mockImplementation( () => ( {
-				data: getActivePersonalPlanDataForType( activePlan ),
-			} ) );
-			const cartChanges = { products: [ getBusinessPlanForInterval( cartPlan ) ] };
-			render( <MyCheckout cartChanges={ cartChanges } />, container );
-			const editOrderButton = await screen.findByLabelText( 'Edit your order' );
-			fireEvent.click( editOrderButton );
-
-			const variantItem = screen
-				.getByText( getVariantItemTextForInterval( expectedVariant ) )
-				.closest( 'label' );
-			expect( within( variantItem ).queryByText( /Save \d+%/ ) ).not.toBeInTheDocument();
-		}
-	);
-
-	it( 'does not render the variant picker if there are no variants after clicking into edit mode', async () => {
-		const cartChanges = { products: [ domainProduct ] };
-		render( <MyCheckout cartChanges={ cartChanges } />, container );
-		const editOrderButton = await screen.findByLabelText( 'Edit your order' );
-		fireEvent.click( editOrderButton );
-
-		expect( screen.queryByText( 'One month' ) ).not.toBeInTheDocument();
-		expect( screen.queryByText( 'One year' ) ).not.toBeInTheDocument();
-		expect( screen.queryByText( 'Two years' ) ).not.toBeInTheDocument();
-	} );
-
-	it.each( [
-		{ activePlan: 'yearly', cartPlan: 'monthly' },
-		{ activePlan: 'monthly', cartPlan: 'yearly' },
-	] )(
-		'does not render the variant picker for a term change from $activePlan to $cartPlan of the current plan',
-		async ( { activePlan, cartPlan } ) => {
-			getPlansBySiteId.mockImplementation( () => ( {
-				data: getActivePersonalPlanDataForType( activePlan ),
-			} ) );
-			const cartChanges = { products: [ getPersonalPlanForInterval( cartPlan ) ] };
-			render( <MyCheckout cartChanges={ cartChanges } />, container );
-			const editOrderButton = await screen.findByLabelText( 'Edit your order' );
-			fireEvent.click( editOrderButton );
-
-			expect( screen.queryByText( 'One month' ) ).not.toBeInTheDocument();
-			expect( screen.queryByText( 'One year' ) ).not.toBeInTheDocument();
-			expect( screen.queryByText( 'Two years' ) ).not.toBeInTheDocument();
-		}
-	);
-
-	it( 'does not render the variant picker for a renewal of the current plan', async () => {
-		const currentPlanRenewal = { ...planWithoutDomain, extra: { purchaseType: 'renewal' } };
-		const cartChanges = { products: [ currentPlanRenewal ] };
-		render( <MyCheckout cartChanges={ cartChanges } />, container );
-		const editOrderButton = await screen.findByLabelText( 'Edit your order' );
-		fireEvent.click( editOrderButton );
-
-		expect( screen.queryByText( 'One month' ) ).not.toBeInTheDocument();
-		expect( screen.queryByText( 'One year' ) ).not.toBeInTheDocument();
-		expect( screen.queryByText( 'Two years' ) ).not.toBeInTheDocument();
 	} );
 
 	it( 'removes a product from the cart after clicking to remove it in edit mode', async () => {
